feat(HW-639): ProviderStore refactoring
