+++ src/helpers/get-rpc-provider.ts
@@ -1,10 +1,10 @@
 import {ethers} from 'ethers';
 
-import {Provider} from '@app/models/provider';
+import {ProviderModel} from '@app/models/provider';
 
 import {EthRpcEndpointAvailability} from './eth-rpc-endpoint-availability';
 
-export async function getRpcProvider(provider: Provider) {
+export async function getRpcProvider(provider: ProviderModel) {
   await EthRpcEndpointAvailability.awaitForInitialization();
   return new ethers.providers.StaticJsonRpcProvider(provider.ethRpcEndpoint, {
     chainId: provider.ethChainId,
