+++ src/models/provider/index.ts
@@ -1,3 +1,5 @@
 export * from './provider';
 export * from './provider-config';
+export * from './provider-config.model';
+export * from './provider.model';
 export * from './provider.types';
